Fixed a bug with npc->HitDir.

+++ src/guys.h
@@ -114,7 +114,7 @@ public:
 	
 	bool immortal;
 	bool noSlide;
-	byte hitdir;
+	signed short hitdir;
 	
 	int getScriptUID();
 	void setScriptUID(int new_id);
